fix tests + removed useNavigate mock
