+++ frontend/tests/components/scan/utils/RecentScans.test.jsx
@@ -12,12 +12,6 @@ import {
 } from "../../../../src/constants/apiURLs";
 
 jest.mock("axios");
-// mock navigate
-const mockedNavigate = jest.fn();
-jest.mock("react-router-dom", () => ({
-  ...jest.requireActual("react-router-dom"),
-  useNavigate: () => mockedNavigate,
-}));
 
 describe("Recent Scans test", () => {
   const recentScansUser = [
@@ -235,8 +229,6 @@ describe("Recent Scans test", () => {
 
       // check redirect to job page
       user.click(firstCard);
-      expect(mockedNavigate).toHaveBeenCalledTimes(1);
-      expect(mockedNavigate).toHaveBeenCalledWith("/jobs/1/visualizer/");
     });
   });
 });
