+++ frontend/tests/mock.js
@@ -91,6 +91,117 @@ export const mockedUseTagsStore = {
   update: () => {},
 };
 
+export const mockedPlaybooks = {
+  TEST_PLAYBOOK_IP: {
+    name: "TEST_PLAYBOOK_IP",
+    type: ["ip"],
+    description: "Test playbook for the IP addresses",
+    disabled: false,
+    runtime_configuration: {
+      analyzers: {},
+      connectors: {},
+      visualizers: {},
+    },
+    analyzers: [],
+    connectors: [],
+    scan_mode: 2,
+    scan_check_time: "02:00:00:00",
+    tags: [
+      {
+        id: 1,
+        label: "test tag",
+        color: "#1655D3",
+      },
+    ],
+    tlp: "CLEAR",
+  },
+  TEST_PLAYBOOK_DOMAIN: {
+    name: "TEST_PLAYBOOK_DOMAIN",
+    type: ["domain"],
+    description: "Test playbook for the domains",
+    disabled: false,
+    runtime_configuration: {
+      analyzers: {},
+      connectors: {},
+      visualizers: {},
+    },
+    analyzers: [],
+    connectors: [],
+    scan_mode: 2,
+    scan_check_time: "02:00:00:00",
+    tags: [],
+    tlp: "CLEAR",
+  },
+  TEST_PLAYBOOK_URL: {
+    name: "TEST_PLAYBOOK_URL",
+    type: ["url"],
+    description: "Test playbook for the URLs",
+    disabled: false,
+    runtime_configuration: {
+      analyzers: {},
+      connectors: {},
+      visualizers: {},
+    },
+    analyzers: [],
+    connectors: [],
+    scan_mode: 1,
+    scan_check_time: null,
+    tags: [],
+    tlp: "AMBER",
+  },
+  TEST_PLAYBOOK_HASH: {
+    name: "TEST_PLAYBOOK_HASH",
+    type: ["hash"],
+    description: "Test playbook for the hashes",
+    disabled: false,
+    runtime_configuration: {
+      analyzers: {},
+      connectors: {},
+      visualizers: {},
+    },
+    analyzers: [],
+    connectors: [],
+    scan_mode: 1,
+    scan_check_time: null,
+    tags: [],
+    tlp: "AMBER",
+  },
+  TEST_PLAYBOOK_FILE: {
+    name: "TEST_PLAYBOOK_FILE",
+    type: ["file"],
+    description: "Test playbook for the files",
+    disabled: false,
+    runtime_configuration: {
+      analyzers: {},
+      connectors: {},
+      visualizers: {},
+    },
+    analyzers: [],
+    connectors: [],
+    scan_mode: 1,
+    scan_check_time: null,
+    tags: [],
+    tlp: "AMBER",
+  },
+  TEST_PLAYBOOK_GENERIC: {
+    name: "TEST_PLAYBOOK_GENERIC",
+    type: ["generic"],
+    description: "Test playbook for the generic observables",
+    disabled: false,
+    runtime_configuration: {
+      analyzers: {},
+      connectors: {},
+      visualizers: {},
+    },
+    analyzers: [],
+    connectors: [],
+    scan_mode: 1,
+    scan_check_time: null,
+    tags: [],
+    tlp: "AMBER",
+  },
+}
+
 export const mockedUsePluginConfigurationStore = {
   analyzersLoading: false,
   connectorsLoading: false,
@@ -141,114 +252,12 @@ export const mockedUsePluginConfigurationStore = {
   visualizers: [],
   ingestors: [],
   playbooks: [
-    {
-      name: "TEST_PLAYBOOK_IP",
-      type: ["ip"],
-      description: "Test playbook for the IP addresses",
-      disabled: false,
-      runtime_configuration: {
-        analyzers: {},
-        connectors: {},
-        visualizers: {},
-      },
-      analyzers: [],
-      connectors: [],
-      scan_mode: 2,
-      scan_check_time: "02:00:00:00",
-      tags: [
-        {
-          id: 1,
-          label: "test tag",
-          color: "#1655D3",
-        },
-      ],
-      tlp: "CLEAR",
-    },
-    {
-      name: "TEST_PLAYBOOK_DOMAIN",
-      type: ["domain"],
-      description: "Test playbook for the domains",
-      disabled: false,
-      runtime_configuration: {
-        analyzers: {},
-        connectors: {},
-        visualizers: {},
-      },
-      analyzers: [],
-      connectors: [],
-      scan_mode: 2,
-      scan_check_time: "02:00:00:00",
-      tags: [],
-      tlp: "CLEAR",
-    },
-    {
-      name: "TEST_PLAYBOOK_URL",
-      type: ["url"],
-      description: "Test playbook for the URLs",
-      disabled: false,
-      runtime_configuration: {
-        analyzers: {},
-        connectors: {},
-        visualizers: {},
-      },
-      analyzers: [],
-      connectors: [],
-      scan_mode: 1,
-      scan_check_time: null,
-      tags: [],
-      tlp: "AMBER",
-    },
-    {
-      name: "TEST_PLAYBOOK_HASH",
-      type: ["hash"],
-      description: "Test playbook for the hashes",
-      disabled: false,
-      runtime_configuration: {
-        analyzers: {},
-        connectors: {},
-        visualizers: {},
-      },
-      analyzers: [],
-      connectors: [],
-      scan_mode: 1,
-      scan_check_time: null,
-      tags: [],
-      tlp: "AMBER",
-    },
-    {
-      name: "TEST_PLAYBOOK_FILE",
-      type: ["file"],
-      description: "Test playbook for the files",
-      disabled: false,
-      runtime_configuration: {
-        analyzers: {},
-        connectors: {},
-        visualizers: {},
-      },
-      analyzers: [],
-      connectors: [],
-      scan_mode: 1,
-      scan_check_time: null,
-      tags: [],
-      tlp: "AMBER",
-    },
-    {
-      name: "TEST_PLAYBOOK_GENERIC",
-      type: ["generic"],
-      description: "Test playbook for the generic observables",
-      disabled: false,
-      runtime_configuration: {
-        analyzers: {},
-        connectors: {},
-        visualizers: {},
-      },
-      analyzers: [],
-      connectors: [],
-      scan_mode: 1,
-      scan_check_time: null,
-      tags: [],
-      tlp: "AMBER",
-    },
+    mockedPlaybooks.TEST_PLAYBOOK_IP,
+    mockedPlaybooks.TEST_PLAYBOOK_URL,
+    mockedPlaybooks.TEST_PLAYBOOK_DOMAIN,
+    mockedPlaybooks.TEST_PLAYBOOK_HASH,
+    mockedPlaybooks.TEST_PLAYBOOK_FILE,
+    mockedPlaybooks.TEST_PLAYBOOK_GENERIC,
   ],
   hydrate: () => {},
   retrieveAnalyzersConfiguration: () => {},
